001, clear frozen status when starts; check and do not download already imported files; paginations for api-records, more errors control

+++ models/data/file.go
@@ -20,7 +20,7 @@ func CreateFileFromPal(palFile *modelsPal.File) *File {
 		Id: 			uuid.NewV4().String(),
 		ExternalId:		palFile.Id,
 		Filename:		palFile.FileName,
-		Size:			int64(palFile.FileSize),
+		Size:			palFile.FileSize,
 		Description:	palFile.Description,
 		Title:			palFile.Title,
 		Fullpath:		palFile.OutFilename,
